Feat: 마이페이지 레이아웃 및 게시글 카드 리스트 UI 구현

+++ backend_core/src/main/java/com/project/plant_parent/repository/FollowRepository.java
@@ -40,11 +40,11 @@ public interface FollowRepository extends JpaRepository<Follow, Long> {
 
     // 팔로잉 수
     // select count(*) from follows where from_member_id = ?
-    int countByFromMember(Member fromMember);
+    long countByFromMember(Member fromMember);
 
     // 팔로워 수
     // select count(*) from follows where to_member_id = ?
-    int countByToMember(Member toMember);
+    long countByToMember(Member toMember);
